added admin controllers and added auth to journal entries

+++ src/main/java/net/engineeringdigest/journalApp/service/impl/JournalServiceImplementation.java
@@ -23,15 +23,15 @@ public class JournalServiceImplementation implements JournalService {
     private final UserService userService;
 
     @Transactional
-    public void createJournalEntry(JournalEntry entry, String userName) {
+    public JournalEntry createJournalEntry(JournalEntry entry, String userName) {
         try {
             User user = userService.findByUserName(userName);
             entry.setDate(LocalDateTime.now());
             JournalEntry saved = journalRepository.save(entry);
             user.getJournalEntries().add(saved);
-            userService.createUser(user);
-        } catch(Exception e) {
-            System.out.println(e);
+            userService.saveUser(user);
+            return saved;
+        } catch (Exception e) {
             throw new RuntimeException("An error occurred when trying to save journal for a user.", e);
         }
     }
@@ -41,28 +41,40 @@ public class JournalServiceImplementation implements JournalService {
         return journalRepository.save(entry);
     }
 
-    public List<JournalEntry> getAllJournalEntries() {
-        return journalRepository.findAll();
-    }
-
     public Optional<JournalEntry> getJournalEntryById(ObjectId id) {
         return journalRepository.findById(String.valueOf(id));
     }
 
-    public void deleteJournalEntryById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(entry -> entry.getId().equals(id));
-        userService.createUser(user);
-        journalRepository.deleteById(String.valueOf(id));
+    public List<JournalEntry> findJournalEntriesByUsername(String userName) {
+        return null;
+    }
+
+    @Transactional
+    public boolean deleteJournalEntryById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(entry -> entry.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalRepository.deleteById(String.valueOf(id));
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred when trying to delete journal entry", e);
+        }
+        return removed;
     }
 
     public JournalEntry updateJournalEntry(ObjectId id, JournalEntry entry) {
         JournalEntry journalEntry = journalRepository
-                .findById(String.valueOf(id)).orElseThrow(() -> new NoSuchElementException("Not found"));
+                .findById(String.valueOf(id))
+                .orElseThrow(() -> new NoSuchElementException("Not found"));
         if (journalEntry != null) {
-            journalEntry.setTitle(!entry.equals("") ? entry.getTitle() : journalEntry.getTitle());
-            journalEntry.setContent(entry.getContent() != null && !entry.equals("") ? entry.getContent() : journalEntry.getContent());
+            journalEntry.setTitle(!entry.getTitle().isEmpty() ? entry.getTitle() : journalEntry.getTitle());
+            journalEntry
+                    .setContent(entry.getContent() != null && !entry.getContent().isEmpty() ? entry.getContent() : journalEntry.getContent());
         }
         return journalEntry;
     }
+
 }
